tests/wpt/tests/webvtt/parsing/file-parsing/tests/comment-in-cue-text.html

<!doctype html>
<!-- DO NOT EDIT! This file and support/comment-in-cue-text.vtt are generated. -->
<!-- See /webvtt/parsing/file-parsing/README.md -->
<meta charset=utf-8>
<title>WebVTT parser test: comment-in-cue-text</title>
<link rel="help" href="https://www.w3.org/TR/webvtt1/#webvtt-comment-block">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test('comment-in-cue-text');
t.step(function(){
    var video = document.createElement('video');
    var track = document.createElement('track');
    assert_true('src' in track, 'track element not supported');
    track.src = 'support/comment-in-cue-text.vtt';
    track['default'] = true;
    track.kind = 'subtitles';
    track.onload = this.step_func(trackLoaded);
    track.onerror = this.step_func(trackError);
    video.appendChild(track);
    document.body.appendChild(video);
});

function trackLoaded(event) {
    var track = event.target;
    var video = track.parentNode;
    var cues = video.textTracks[0].cues;
    {
assert_equals(cues.length, 2);

assert_equals(cues[0].text, 'NOTE text');
assert_equals(cues[0].startTime, 0);
assert_equals(cues[0].endTime, 1);

assert_equals(cues[1].text, 'NOTE text\nNOTE text2');
assert_equals(cues[1].startTime, 1);
assert_equals(cues[1].endTime, 2);

    }
    this.done();
}

function trackError(e) {
    assert_unreached('got unexpected error event');
}
</script>
